# perf: optimize typing screen transition performance for issue #244