--- a/src/extractor/code_chunk_extractor.rs
+++ b/src/extractor/code_chunk_extractor.rs
@@ -21,7 +21,7 @@ impl CodeChunkExtractor {
     pub fn extract_chunks_from_files_with_progress<P: ProgressReporter + ?Sized>(
         &mut self,
         files_to_process: Vec<(std::path::PathBuf, Box<dyn Language>)>,
-        _options: &ExtractionOptions,
+        options: &ExtractionOptions,
         progress: &P,
     ) -> Result<Vec<CodeChunk>> {
         let total_files = files_to_process.len();
@@ -29,7 +29,7 @@ impl CodeChunkExtractor {
         // Sort files by size (largest first) for better loading progress perception
         // Cache metadata to avoid repeated filesystem calls
         let mut files_with_sizes: Vec<_> = files_to_process
-            .into_iter()
+            .into_par_iter()
             .map(|(path, lang)| {
                 let size = std::fs::metadata(&path).map(|m| m.len()).unwrap_or(0);
                 (path, lang, size)
@@ -38,9 +38,10 @@ impl CodeChunkExtractor {
 
         files_with_sizes.sort_by(|a, b| b.2.cmp(&a.2)); // Sort by size (largest first)
 
+        // Convert back but keep size info for logging
         let files_to_process: Vec<_> = files_with_sizes
-            .into_iter()
-            .map(|(path, lang, _size)| (path, lang))
+            .into_par_iter()
+            .map(|(path, lang, _size)| (path, lang, _size))
             .collect();
 
         // Initialize extracting progress from 0
@@ -49,26 +50,27 @@ impl CodeChunkExtractor {
         // Process all files in parallel with atomic progress tracking
         let processed = Arc::new(AtomicUsize::new(0));
 
+        // Use into_par_iter for better work-stealing efficiency
         let all_chunks: Vec<CodeChunk> = files_to_process
             .into_par_iter()
-            .flat_map(|(path, language)| {
-                let result = Self::extract_from_file_static(&path, language.as_ref(), _options);
+            .flat_map(|(path, language, _size)| {
+                let result = Self::extract_from_file_static(&path, language.as_ref(), options);
 
                 // Update progress atomically
                 let current = processed.fetch_add(1, Ordering::Relaxed) + 1;
 
-                // Update progress display every 10 files or on completion
-                if current % 10 == 0 || current == total_files {
+                // Dynamic progress update frequency: every 10 files, but every 1 file when >99%
+                let progress_threshold = if current as f64 / total_files as f64 > 0.99 {
+                    1 // Update every file when >99%
+                } else {
+                    10 // Update every 10 files normally
+                };
+
+                if current % progress_threshold == 0 || current == total_files {
                     progress.set_file_counts(StepType::Extracting, current, total_files, None);
                 }
 
-                match result {
-                    Ok(file_chunks) => file_chunks,
-                    Err(e) => {
-                        eprintln!("Warning: Failed to extract from file {:?}: {}", path, e);
-                        Vec::new()
-                    }
-                }
+                result.unwrap_or(Vec::new())
             })
             .collect();
 
@@ -82,8 +84,21 @@ impl CodeChunkExtractor {
     fn extract_from_file_static(
         file_path: &Path,
         language: &dyn Language,
-        _options: &ExtractionOptions,
+        options: &ExtractionOptions,
     ) -> Result<Vec<CodeChunk>> {
+        if let Ok(metadata) = std::fs::metadata(file_path) {
+            let file_size = metadata.len();
+            if file_size > options.max_file_size_bytes {
+                log::warn!(
+                    "Skipping large file: {:?} ({}MB > {}MB limit)",
+                    file_path,
+                    file_size / (1024 * 1024),
+                    options.max_file_size_bytes / (1024 * 1024)
+                );
+                return Ok(Vec::new());
+            }
+        }
+        
         CommonExtractor::extract_from_file(file_path, language.name())
     }
 }