… Cargo.lock
… Cargo.toml
… src/extractor/core/extractor.rs
… src/extractor/models/options.rs
… src/extractor/repository_extractor.rs
… src/game/game_data.rs
… src/game/models/loading_steps/cache_check_step.rs
… src/game/models/loading_steps/extracting_step.rs
… src/game/models/loading_steps/finalizing_step.rs
… src/game/models/loading_steps/step_manager.rs
… src/game/screen_manager.rs
… src/game/screens/typing_screen.rs
… src/game/session_manager.rs
… src/game/stage_repository.rs
… src/game/typing_core.rs
… tests/integration/comment_processing_tests.rs
… tests/integration/indent_treesitter_tests.rs
… tests/integration/languages/c/extractor.rs
… tests/integration/languages/cpp/extractor.rs
… tests/integration/languages/csharp/extractor.rs
… tests/integration/languages/dart/extractor.rs
… tests/integration/languages/go/extractor.rs
… tests/integration/languages/haskell/extractor.rs
… tests/integration/languages/java/extractor.rs
… tests/integration/languages/javascript/extractor.rs
… tests/integration/languages/kotlin/extractor.rs
… tests/integration/languages/php/extractor.rs
… tests/integration/languages/python/extractor.rs
… tests/integration/languages/ruby/extractor.rs
… tests/integration/languages/rust/extractor.rs